+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added dripping upgrade that will fill fluid drawers with lava
 * Added water source upgrade that will fill fluid drawers with water
 * Fix rebuild voxel shape cost too much time, close #314
+* fix: #310 remove instantiate of java.awt.Color
 
 # VERSION 1.3.2
 * Fix fluid drawers not being accessible through controllers, closes #301